services/api/requirements.txt
fastapi
uvicorn[standard]
httpx
redis>=5
pydantic
PyJWT